feat: introduce generic readers instead of the StreamReaderStrategy.
This has several advantages:
- no instanceof
- no handling of non-intended use required
- only one interface with no restrictions on type T

Also provides a helper/factory function for the construction of available readers.

Downsides:
- Deprecates some existing API for the next major version
- Generics required in reading code (e.g. CrateReader<String> in future)
