+++ src/main/java/edu/kit/datamanager/ro_crate/reader/ReaderStrategy.java
@@ -1,18 +1,14 @@
 package edu.kit.datamanager.ro_crate.reader;
 
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import java.io.File;
-
 /**
- * Interface for the strategy fo the reader class.
+ * Interface for the strategy for the reader class.
  * This should be implemented if additional strategies are to be build.
  * (e.g., reading from a gzip)
  *
  * @author Nikola Tzotchev on 9.2.2022 г.
  * @version 1
+ *
+ * @deprecated Use {@link GenericReaderStrategy} instead.
  */
-public interface ReaderStrategy {
-  ObjectNode readMetadataJson(String location);
-
-  File readContent(String location);
-}
+@Deprecated(since = "2.1.0", forRemoval = true)
+public interface ReaderStrategy extends GenericReaderStrategy<String> {}
