+++ src/main/java/edu/kit/datamanager/ro_crate/reader/ZipReader.java
@@ -1,17 +1,6 @@
 package edu.kit.datamanager.ro_crate.reader;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-import edu.kit.datamanager.ro_crate.objectmapper.MyObjectMapper;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.UUID;
 import java.nio.file.Path;
-
-import net.lingala.zip4j.ZipFile;
-import org.apache.commons.io.FileUtils;
 
 /**
  * A ReaderStrategy implementation which reads from ZipFiles.
@@ -28,17 +17,18 @@ import org.apache.commons.io.FileUtils;
  * folder after extraction. Use RoCrateWriter to export it so some
  * persistent location and possibly read it from there, if required. Or use
  * the ZipWriter to write it back to its source.
+ *
+ * @deprecated Use {@link ZipStrategy} instead.
  */
-public class ZipReader implements ReaderStrategy {
-
-  protected final String ID = UUID.randomUUID().toString();
-  protected Path temporaryFolder = Path.of(String.format("./.tmp/ro-crate-java/zipReader/%s/", ID));
-  protected boolean isExtracted = false;
+@Deprecated(since = "2.1.0", forRemoval = true)
+public class ZipReader extends ZipStrategy {
 
   /**
    * Crates a ZipReader with the default configuration as described in the class documentation.
    */
-  public ZipReader() {}
+  public ZipReader() {
+    super();
+  }
 
   /**
    * Creates a ZipReader which will extract the contents temporary
@@ -52,77 +42,6 @@ public class ZipReader implements ReaderStrategy {
    *                              will have UUIDs as their names.
    */
   public ZipReader(Path folderPath, boolean shallAddUuidSubfolder) {
-    if (shallAddUuidSubfolder) {
-      this.temporaryFolder = folderPath.resolve(ID);
-    } else {
-      this.temporaryFolder = folderPath;
-    }
-  }
-
-  /**
-   * @return the identifier which may be used as the name for a subfolder in the temporary directory.
-   */
-  public String getID() {
-    return ID;
-  }
-
-  /**
-   * @return the folder (considered temporary) where the zipped crate will be or has been extracted to.
-   */
-  public Path getTemporaryFolder() {
-    return temporaryFolder;
-  }
-
-  /**
-   * @return whether the crate has already been extracted into the temporary folder.
-   */
-  public boolean isExtracted() {
-    return isExtracted;
-  }
-
-  private void readCrate(String location) {
-    try {
-      File folder = temporaryFolder.toFile();
-      // ensure the directory is clean
-      if (folder.isDirectory()) {
-        FileUtils.cleanDirectory(folder);
-      } else if (folder.isFile()) {
-        FileUtils.delete(folder);
-      }
-      // extract
-      try (ZipFile zf = new ZipFile(location)) {
-        zf.extractAll(temporaryFolder.toAbsolutePath().toString());
-        this.isExtracted = true;
-      }
-      // register deletion on exit
-      FileUtils.forceDeleteOnExit(folder);
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-  }
-
-  @Override
-  public ObjectNode readMetadataJson(String location) {
-    if (!isExtracted) {
-      this.readCrate(location);
-    }
-
-    ObjectMapper objectMapper = MyObjectMapper.getMapper();
-    File jsonMetadata = temporaryFolder.resolve("ro-crate-metadata.json").toFile();
-    
-    try {
-      return objectMapper.readTree(jsonMetadata).deepCopy();
-    } catch (IOException e) {
-      e.printStackTrace();
-      return null;
-    }
-  }
-
-  @Override
-  public File readContent(String location) {
-    if (!isExtracted) {
-      this.readCrate(location);
-    }
-    return temporaryFolder.toFile();
+    super(folderPath, shallAddUuidSubfolder);
   }
 }
